update README.md, remove context, add hooks for toggle audio and microphone
