+++ example/App.tsx
@@ -3,15 +3,15 @@ import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import Toast from 'react-native-toast-message';
 
 import AppNavigator from './navigators/AppNavigator';
-import AppProvider from './providers/AppProvider';
+import { FishjamContextProvider } from '@fishjam-dev/react-native-client';
 
 function App(): React.JSX.Element {
   return (
     <>
       <GestureHandlerRootView>
-        <AppProvider>
+        <FishjamContextProvider>
           <AppNavigator />
-        </AppProvider>
+        </FishjamContextProvider>
       </GestureHandlerRootView>
       <Toast />
     </>
