--- example/providers/AppProvider.tsx
@@ -1,15 +1,0 @@
-import { FishjamContextProvider } from '@fishjam-dev/react-native-client';
-import React from 'react';
-
-import { FishjamExampleContextProvider } from '../contexts/FishjamExampleContext';
-import type { AppParentNode } from '../types/AppParentNode';
-
-const AppProvider: React.FC<AppParentNode> = ({ children }) => {
-  return (
-    <FishjamContextProvider>
-      <FishjamExampleContextProvider>{children}</FishjamExampleContextProvider>
-    </FishjamContextProvider>
-  );
-};
-
-export default AppProvider;
